pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = "A lightweight API Gateway built with FastAPI, featuring health checks and Prometheus metrics."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"

# Core framework
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }

# Request handling and authentication
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

# Monitoring & Metrics
prometheus-client = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
